add event UI + add order management

+++ src/components/CartItem/CartItem.js
@@ -23,7 +23,23 @@ const CartItem = ({
   const [itemQuantity, setItemQuantity] = useState(quantity)
 
   const removeItem = () => {
-    dispatch(removeProductToCart(id))
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this imaginary file!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          dispatch(removeProductToCart(id))
+          swal("Poof! Your imaginary file has been deleted!", {
+            icon: "success",
+          });
+        } else {
+          swal("Your imaginary file is safe!");
+        }
+      });
   }
 
   const handleQuantityChange = (e) => {
